started propmanIT, added sorting to findalls, bugfix and slight edits to tagmanIT

+++ src/test/java/gov/bnl/channelfinder/PropertyManagerIT.java
@@ -1,11 +1,13 @@
 package gov.bnl.channelfinder;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 import org.junit.After;
@@ -38,7 +40,7 @@ public class PropertyManagerIT {
     ChannelRepository channelRepository;
 
     // set up
-    XmlProperty testProperty = new XmlProperty("testProperty","testOwner");
+    XmlProperty testProperty0 = new XmlProperty("testProperty0","testOwner");
     XmlProperty testProperty1 = new XmlProperty("testProperty1","testOwner1");
     XmlProperty testProperty2 = new XmlProperty("testProperty2","testOwner2");   
     XmlProperty testPropertyC = new XmlProperty("testPropertyC","testOwnerC");    
@@ -51,74 +53,130 @@ public class PropertyManagerIT {
     /**
      * list all properties
      */
-    @Test // might actually work but messed up by full database
+    @Test 
     public void listXmlProperties() {
-        testChannels.forEach(chan -> chan.addProperty(testPropertyC));
-        testPropertyC.setChannels(testChannels);
-        List<XmlProperty> testProperties = Arrays.asList(testProperty,testPropertyC);
-        Iterable<XmlProperty> createdProperty = propertyManager.create(testProperties);
-
+        XmlProperty testProperty0 = new XmlProperty("testProperty0","testOwner");
+        XmlProperty testProperty1 = new XmlProperty("testProperty1","testOwner1");
+        testProperty1.setChannels(Arrays.asList(
+                new XmlChannel(testChannels.get(0).getName(),testChannels.get(0).getOwner(),Arrays.asList(new XmlProperty(testProperty1.getName(),testProperty1.getOwner(),"value")),new ArrayList<XmlTag>())));
+        List<XmlProperty> testProperties = Arrays.asList(testProperty0,testProperty1);
+        cleanupTestProperties = testProperties;
+        
+        Iterable<XmlProperty> createdProperties = propertyManager.create(testProperties);
         Iterable<XmlProperty> propertyList = propertyManager.list();
-        for(XmlProperty property: createdProperty) {
+        for(XmlProperty property: createdProperties) {
             property.setChannels(new ArrayList<XmlChannel>());
         }
         // verify the properties were listed as expected
-        assertEquals("Failed to list all properties",createdProperty,propertyList);                
+        assertEquals("Failed to list all properties",createdProperties,propertyList);                
     }
 
     /**
      * read a single property
+     * test the "withChannels" flag
      */
     @Test
     public void readXmlProperty() {
-        testPropertyC.setChannels(testChannels);
-        XmlProperty createdProperty = propertyManager.create(testProperty.getName(),testProperty);
-        XmlProperty createdProperty1 = propertyManager.create(testPropertyC.getName(),testPropertyC);
+        XmlProperty testProperty0 = new XmlProperty("testProperty0","testOwner");
+        XmlProperty testProperty1 = new XmlProperty("testProperty1","testOwner1");
+        testProperty1.setChannels(Arrays.asList(
+                new XmlChannel(testChannels.get(0).getName(),testChannels.get(0).getOwner(),Arrays.asList(new XmlProperty(testProperty1.getName(),testProperty1.getOwner(),"value")),new ArrayList<XmlTag>())));
+        cleanupTestProperties = Arrays.asList(testProperty0,testProperty1);
+        
+        XmlProperty createdProperty0 = propertyManager.create(testProperty0.getName(),testProperty0);
+        XmlProperty createdProperty1 = propertyManager.create(testProperty1.getName(),testProperty1);
 
-        XmlProperty readProperty = propertyManager.read(createdProperty.getName(), false);
+        // verify the created properties are read as expected
+        // retrieve the testProperty0 without channels
+        XmlProperty retrievedProperty = propertyManager.read(createdProperty0.getName(), false);
+        assertEquals("Failed to read the property",createdProperty0,retrievedProperty);        
+        // retrieve the testProperty0 with channels
+        retrievedProperty = propertyManager.read(createdProperty0.getName(), true);
+        assertEquals("Failed to read the property w/ channels",createdProperty0,retrievedProperty);
+
+        retrievedProperty = propertyManager.read(createdProperty1.getName(), false);
+        testProperty1.setChannels(new ArrayList<XmlChannel>());
         // verify the property was read as expected
-        assertEquals("Failed to read the property",createdProperty,readProperty);        
+        assertEquals("Failed to read the property",testProperty1,retrievedProperty);
 
-        readProperty = propertyManager.read(createdProperty.getName(), true);
+        retrievedProperty = propertyManager.read(createdProperty1.getName(), true);
         // verify the property was read as expected
-        assertEquals("Failed to read the property w/ channels",createdProperty,readProperty);
-
-        readProperty = propertyManager.read(createdProperty1.getName(), false);
-        testPropertyC.setChannels(new ArrayList<XmlChannel>());
-        // verify the property was read as expected
-        assertEquals("Failed to read the property",testPropertyC,readProperty);
-
-        readProperty = propertyManager.read(createdProperty1.getName(), true);
-        // verify the property was read as expected
-        assertEquals("Failed to read the property w/ channels",createdProperty1,readProperty);
-
-        try {
-            // verify the property failed to be read, as expected
-            readProperty = propertyManager.read("fakeProperty", false);
-            assertTrue("Failed to throw an error",false);
-        } catch(ResponseStatusException e) {
-            assertTrue(true);
-        }
-
-        try {
-            // verify the property failed to be read, as expected
-            readProperty = propertyManager.read("fakeProperty", true);
-            assertTrue("Failed to throw an error",false);
-        } catch(ResponseStatusException e) {
-            assertTrue(true);
-        }
+        assertEquals("Failed to read the property w/ channels",createdProperty1,retrievedProperty);
+    }
+    
+    /**
+     * attempt to read a single non existent property
+     */
+    @Test(expected = ResponseStatusException.class)
+    public void readNonExistingXmlProperty() {
+        // verify the tag failed to be read, as expected
+        propertyManager.read("fakeProperty", false);
     }
 
+    /**
+     * attempt to read a single non existent property with channels
+     */
+    @Test(expected = ResponseStatusException.class)
+    public void readNonExistingXmlProperty2() {
+        // verify the tag failed to be read, as expected
+        propertyManager.read("fakeProperty", true);
+    }
+    
+    /**
+     * create a simple property
+     */
+    @Test
+    public void createXmlProperty() {
+        XmlProperty testProperty0 = new XmlProperty("testProperty0","testOwner");
+        cleanupTestProperties = Arrays.asList(testProperty0);
+
+        // Create a simple property
+        XmlProperty createdProperty = propertyManager.create(testProperty0.getName(), testProperty0);
+        assertEquals("Failed to create the tag", testProperty0, createdProperty);
+
+//        XmlTag createdTag1 = tagManager.create("fakeTag", copy(testTag1));
+//        // verify the tag was created as expected
+//        assertEquals("Failed to create the tag",testTag1,createdTag1);
+
+        // Update the test property with a new owner
+        XmlProperty updatedTestProperty0 = new XmlProperty("testProperty0", "updateTestOwner");
+        createdProperty = propertyManager.create(testProperty0.getName(), updatedTestProperty0);
+        assertEquals("Failed to create the tag", updatedTestProperty0, createdProperty);
+    }
+    
+    /**
+     * Rename a simple property
+     */
+    @Test
+    public void renameTag() {
+        XmlProperty testProperty0 = new XmlProperty("testProperty0","testOwner");
+        XmlProperty testProperty1 = new XmlProperty("testProperty1","testOwner1");
+        cleanupTestProperties = Arrays.asList(testProperty0,testProperty1);
+        
+        XmlProperty createdProperty = propertyManager.create(testProperty0.getName(), testProperty0);
+        createdProperty = propertyManager.create(testProperty0.getName(), testProperty1);
+        // verify that the old property "testProperty0" was replaced with the new "testProperty1"
+        assertEquals("Failed to create the property", testProperty1, createdProperty);
+        // verify that the old property is no longer present
+        assertFalse("Failed to replace the old property", tagRepository.existsById(testProperty0.getName()));
+    }
+    
+    
+    
+    
+    
+    
+    
+    
     // Helper operations to create and clean up the resources needed for successful
     // testing of the PropertyManager operations
 
     private final List<XmlChannel> testChannels = Arrays.asList(
-            new XmlChannel("testChannel", "testOwner"),
+            new XmlChannel("testChannel0", "testOwner"),
             new XmlChannel("testChannel1", "testOwner"));
 
-    private final List<XmlProperty> allTestProperties = Arrays.asList(
-            testProperty,testProperty1,testProperty2,testPropertyC,testPropertyC1,testPropertyC2,updateTestProperty,updateTestPropertyC,updateTestPropertyC1);
-
+    private List<XmlProperty> cleanupTestProperties = Collections.emptyList();
+    
     @Before
     public void setup() {
         channelRepository.indexAll(testChannels);
@@ -130,12 +188,14 @@ public class PropertyManagerIT {
         testChannels.forEach(channel -> { 
             try {
                 channelRepository.deleteById(channel.getName());
-            } catch (Exception e) {}
+            } catch (Exception e) {      
+                System.out.println("Failed to clean up channel: " + channel.getName());
+            }
         });
-        allTestProperties.forEach(property -> {
-            try {
+        cleanupTestProperties.forEach(property -> {
+            if (propertyRepository.existsById(property.getName())) {
                 propertyRepository.deleteById(property.getName());
-            } catch (Exception e) {}
+            } 
         });
     }
